ml_service/requirements.txt
flask==2.3.3
flask-cors==4.0.0
pandas==2.0.3
numpy==1.24.3
scikit-learn==1.3.0
joblib==1.3.2
requests==2.31.0
xgboost==1.7.6
lightgbm==4.0.0
tensorflow==2.13.0
prophet==1.1.4
PyJWT==2.8.0
